Merged PR 41601: EZRX- 1659 Tender Contract using remote data source
Please use the following to test -

Sales org : 3050
Customer code : 30282241

1. Go to create order and select material 21129643
By default, Tender contract with reason '730' will be selected.

2. Now select material 21223505
By default, the first item - No contract will be selected.

Expected behaviour -
If the contract list contains a 730 contract, then it is selected by default. Otherwise the first item is selected by default.

Future scope for next ticket -
If a contract 730 item is added to the cart, only items with contract 730 can be added to the cart.