Merged PR 45356: EZRX-2246 Add Tender Contract to saved order
https://zuelligpharma.atlassian.net/browse/EZRX-2246

Sales org: 3050
Customer code: 30282241
Material: 21129643

Select the tender contract and add the material to the cart and save the order.
Go to saved order and you can see the order with tender contract.
You can add the order to the cart. Tender contract is selected and added to the cart.