Merged PR 53057: Fixbug/EZRX-3175-1010-order-template-detail-page-incorrect-display-price-when-change-sales-org
# Root cause:

Since `MaterialPriceDetailBloc` is global access, both `OrderTemplatePage`, `SavedOrderPage`, and `FavoriteTab` use it for fetching prices.

 When we quickly change SalesOrg/CustomerCode/ShipToCode and then access an `OrderTemplateDetailPage`, `MaterialPriceDetailBloc` is used for fetching the price for both order template item and favorite item at the same time so the rotation flow of `isFetching` and `isValidating` statuses are mixed-up.

#Solution:
Provide `FavoriteTab` its own `MaterialPriceDetailBloc` because its fetching price process is triggered automatically when we change SalesOrg/CustomerCode/ShipToCode so it won't affect another screen.

#Why does this PR not reinforce the test case?
We can't reproduce this case in the widget test because whenever we test a widget, we only pump one screen at one time.

This case only happens when we have both `FavoriteTab` and `OrderTemplateDetailPage` on the screen