Merged PR 59963: feat/EZRX-5767 : KH Regression test case refinement after latest changes
### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-5767

### Root cause
- Announcement icon & cart page back

### Solution
- need to close announcement icon before submit order

### Test account
- username: khclientuser/St@ysafe01
- cmd:
   make reset_kh_salesorg_config
   make run_kh_client_test
   make run_kh_external_test

### Reminder
- if you code have significant change please run integration test `make run_sg_client_test` to check if you break any test cases

Description:
Announcement icon comes during order creation, as of now, it's visible in material details page and cart page, need to close announcement as it is not a part of ordering and may effect our script (scrolling issue & visibility of finder). So I added in material Description, once it is closed should be appear again.
Also I can see the cart page back button clicked twice, need to call only one, sometimes, it fail sometimes it's not, because of we have tester.tap() methods where it show warning instead of fail if the tap finder is missed.