Merged PR 79306: Feature/EZRX-11200 : Add navigation to payment details page on clever tap notification list for payment
### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-11200

**Screen Recording** -
[Screen Recording 2023-09-23 at 11.19.58 AM.mov](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/79306/attachments/Screen%20Recording%202023-09-23%20at%2011.19.58%20AM.mov)

**Description** -
we need to call two concurrent api [**customerPayment**, **paymentItems** Endpoints], because we need to show some data on details page that we are not getting from paymentItems EndPoint.
So we got two parameter from **Notification** api - salesDocument & PaymentNumber, that we have to pass to paymentItems EndPoint.

**Why we not calling api concurrent**
When we pass data to **customerPayment** & **paymentItems** api's concurrently, I observe the salesDocument as PaymentId & PaymentNumber as paymentBatchAdditionalInfo, not working. PaymentBatchAdditionalInfo is the required fields to get response of **paymentItems** api. But paymentBatchAdditionalInfo we get from **customerPayment** Api. so we have one value paymentBatchAdditionalInfo as dependency to our **paymentItems** Api. In future, we need to ask backend Devs to change **Notification** api of paymentNumber to paymentBatchAdditionalInfo fields or vice versa, because paymentNumber is not needed anymore to get payment details, the required field is paymentBatchAdditionalInfo.

**In This PR** -
1. Dependent Api call - First Api call **customerPayment** Api, if paymentBatchAdditionalInfo is not valid. Then second Api call - **paymentItems** inside PaymentDetailsBloc.
2. UI Structure change {Too badly Implemented} Tried to refactor somehow but I need to refactor more on separate ticket, as this PR 40-45 files changes as of now.

**Need to refactor**
1. Update csv file, all text is missing on csv
2. call initial event to update the common bloc on state.
3. UI can be refactor.