Merged PR 79695: Implementation/EZRX-11744-mobile-v-3-configuration-enable-zdp-5-discount-implementation
### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-11744

### TDD PR url
-

### Root cause
-
First, the price was not updated when I navigate to the product search page from product details page and again came back to product details page.
root cause:  we escape the materials which had price previously in the material price bloc. For this the price is not updated.
Second, if we change any price in the cart then it was reflected to the product details page or vice versa.
root cause:  we use same state variable for cart product and material listing. Due to this the overriding was occurred.

### Solution
- First, added the event before the navigation to the product details page.
Second, added the event after the  the pop function from cart page.

### Test account
- environment:
- username: rootadmin
-sales org: 3070
-Customer: 0030264519
-Material: 21129185

### Output screenshot
-[Screen Recording 2023-10-19 at 6.59.01 PM.mov](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/79695/attachments/Screen%20Recording%202023-10-19%20at%206.59.01%E2%80%AFPM.mov)
[Screen Recording 2023-10-19 at 6.58.21 PM (2).mov](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/79695/attachments/Screen%20Recording%202023-10-19%20at%206.58.21%E2%80%AFPM%20(2).mov)

### Reminder
- if you code have significant change please run integration test `make run_sg_client_test` to check if you break any test cases

EZRX-11744-mobile-v-3-configuration-enable-zdp-5-discount-implementation

EZRX-11744-mobile-v-3-configuration-enable-zdp-5-discount-implementation